refactor: LottoNumber 클래스를 domain의 lotto 패키지로 이동
