+++ src/test/kotlin/domain/lotto/LottoNumberTest.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.lotto
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
